requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycparser==2.21
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
